requirements.txt
streamlit
langchain-core
langchain-qdrant
qdrant-client
langchain_huggingface
python-dotenv
langchain-openai